# Optimizing (de)serialization bounds-checks