--- a/projects/RabbitMQ.Client/client/framing/BasicConsume.cs
+++ b/projects/RabbitMQ.Client/client/framing/BasicConsume.cs
@@ -31,6 +31,7 @@
 
 using System;
 using System.Collections.Generic;
+
 using RabbitMQ.Client.client.framing;
 using RabbitMQ.Client.Impl;
 
@@ -66,10 +67,10 @@ namespace RabbitMQ.Client.Framing.Impl
         public BasicConsume(ReadOnlySpan<byte> span)
         {
             int offset = 2;
-            offset += WireFormatting.ReadShortstr(span, out _queue);
-            offset += WireFormatting.ReadShortstr(span, out _consumerTag);
-            offset += WireFormatting.ReadBits(span.Slice(offset), out _noLocal, out _noAck, out _exclusive, out _nowait);
-            WireFormatting.ReadDictionary(span.Slice(offset), out var tmpDictionary);
+            offset += WireFormatting.ReadShortstr(span, offset, out _queue);
+            offset += WireFormatting.ReadShortstr(span, offset, out _consumerTag);
+            offset += WireFormatting.ReadBits(span, offset, out _noLocal, out _noAck, out _exclusive, out _nowait);
+            WireFormatting.ReadDictionary(span, offset, out var tmpDictionary);
             _arguments = tmpDictionary;
         }
 
@@ -79,11 +80,11 @@ namespace RabbitMQ.Client.Framing.Impl
 
         public override int WriteArgumentsTo(Span<byte> span)
         {
-            int offset = WireFormatting.WriteShort(span, default);
-            offset += WireFormatting.WriteShortstr(span.Slice(offset), _queue);
-            offset += WireFormatting.WriteShortstr(span.Slice(offset), _consumerTag);
-            offset += WireFormatting.WriteBits(span.Slice(offset), _noLocal, _noAck, _exclusive, _nowait);
-            return offset + WireFormatting.WriteTable(span.Slice(offset), _arguments);
+            int offset = WireFormatting.WriteShort(span, 0, default);
+            offset += WireFormatting.WriteShortstr(span, offset, _queue);
+            offset += WireFormatting.WriteShortstr(span, offset, _consumerTag);
+            offset += WireFormatting.WriteBits(ref span[offset], _noLocal, _noAck, _exclusive, _nowait);
+            return offset + WireFormatting.WriteTable(span, offset, _arguments);
         }
 
         public override int GetRequiredBufferSize()