--- a/projects/RabbitMQ.Client/client/impl/MethodBase.cs
+++ b/projects/RabbitMQ.Client/client/impl/MethodBase.cs
@@ -30,6 +30,7 @@
 //---------------------------------------------------------------------------
 
 using System;
+
 using RabbitMQ.Client.client.framing;
 
 namespace RabbitMQ.Client.Impl