… projects/Benchmarks/WireFormatting/DataTypeSerialization.cs
… projects/Benchmarks/WireFormatting/MethodFraming.cs
… projects/Benchmarks/WireFormatting/MethodSerialization.cs
… projects/Benchmarks/WireFormatting/PrimitivesSerialization.cs
… projects/RabbitMQ.Client/client/framing/BasicReject.cs
… projects/RabbitMQ.Client/client/framing/BasicReturn.cs
… projects/RabbitMQ.Client/client/framing/ChannelClose.cs
… projects/RabbitMQ.Client/client/framing/ChannelFlow.cs
… projects/RabbitMQ.Client/client/framing/ChannelFlowOk.cs
… projects/RabbitMQ.Client/client/framing/ConfirmSelect.cs
… projects/RabbitMQ.Client/client/framing/ConnectionBlocked.cs
… projects/RabbitMQ.Client/client/framing/ConnectionClose.cs
… projects/RabbitMQ.Client/client/framing/ConnectionOpen.cs
… projects/RabbitMQ.Client/client/framing/ConnectionSecure.cs
… projects/RabbitMQ.Client/client/framing/ConnectionSecureOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionStart.cs
… projects/RabbitMQ.Client/client/framing/ConnectionStartOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionTune.cs
… projects/RabbitMQ.Client/client/framing/ConnectionTuneOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionUpdateSecret.cs
… projects/RabbitMQ.Client/client/framing/ExchangeBind.cs
… projects/RabbitMQ.Client/client/framing/ExchangeDeclare.cs
… projects/RabbitMQ.Client/client/framing/ExchangeDelete.cs
… projects/RabbitMQ.Client/client/framing/ExchangeUnbind.cs
… projects/RabbitMQ.Client/client/framing/Model.cs
… projects/RabbitMQ.Client/client/framing/Protocol.cs
… projects/RabbitMQ.Client/client/framing/QueueBind.cs
… projects/RabbitMQ.Client/client/framing/QueueDeclare.cs
… projects/RabbitMQ.Client/client/framing/QueueDeclareOk.cs
… projects/RabbitMQ.Client/client/framing/QueueDelete.cs