## Proposed Changes

This PR is not pretty to look at, most of it is ceremony around changes in the WireFormatting class. It does some optimizations to (de)serialization by getting rid of bounds checks and unnecessary arithmetic operations during the serialization phase. There are more improvements that can be made during framing but this results in a hefty net gain for all framing operations I've tested so far.

## Types of Changes

- [ ] Bug fix (non-breaking change which fixes issue #NNNN)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause an observable behavior change in existing systems)
- [ ] Documentation improvements (corrections, new content, etc)
- [ ] Cosmetic change (whitespace, formatting, etc)

## Checklist

- [X] I have read the `CONTRIBUTING.md` document
- [X] I have signed the CA (see https://cla.pivotal.io/sign/rabbitmq)
- [X] All tests pass locally with my changes
- [X] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] Any dependent changes have been merged and published in related repositories

## Further Comments

This PR is not very pretty to look at, but it does result in hefty improvements to command framing both in the deserialization and serialization phase, so it should benefit all operations made within the client.