@stebet 2021-03-09
### .NET Framework 4.8 diff
| Slower                                                                 | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntGetSize              |      1.44 |             1.33 |             1.91 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueGetSize |      1.35 |             9.33 |            12.56 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWriteEmpty         |      1.29 |            22.55 |            29.06 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayReadEmpty          |      1.10 |            18.06 |            19.95 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableGetSizeEmpty       |      1.05 |             8.83 |             9.25 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.StringGetSize           |      1.04 |            46.29 |            47.96 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictRead                |      1.02 |            35.89 |            36.56 |         |

| Faster                                                                     | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrGetSizeEmpty    |      4.85 |             1.97 |             0.41 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayGetSizeEmpty           |      3.87 |             2.19 |             0.57 | bimodal |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictGetSize                 |      2.41 |            15.49 |             6.42 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckWrite                           |      1.80 |            21.22 |            11.77 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckRead                            |      1.71 |            22.14 |            12.93 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrReadEmpty       |      1.70 |            19.03 |            11.17 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseWrite                   |      1.61 |            33.67 |            20.96 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseRead                    |      1.57 |            27.19 |            17.33 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicAck.BasicAckWrite                    |      1.49 |            59.47 |            39.82 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayGetSizePopulated       |      1.48 |            43.98 |            29.66 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.StringRead                  |      1.44 |            44.28 |            30.81 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayRead                   |      1.41 |            26.25 |            18.64 |         |
| RabbitMQ.Benchmarks.MethodFramingChannelClose.ChannelCloseWrite            |      1.28 |            70.35 |            54.91 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishMemoryWrite             |      1.27 |            46.75 |            36.90 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueRead        |      1.26 |            39.56 |            31.44 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishMemoryWrite      |      1.26 |           133.65 |           106.38 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableGetSizePopulated       |      1.22 |           260.91 |           213.33 |         |
| RabbitMQ.Benchmarks.MethodBasicProperties.BasicPropertiesRead              |      1.21 |           103.40 |            85.38 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrWriteEmpty      |      1.16 |            12.35 |            10.62 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayGetSize                |      1.16 |             4.95 |             4.28 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayReadPopulated          |      1.15 |           138.70 |           120.44 |         |
| RabbitMQ.Benchmarks.PrimitivesLong.LongWrite(value: 12345)                 |      1.14 |            11.42 |            10.03 |         |
| RabbitMQ.Benchmarks.PrimitivesShort.ShortWrite(value: 12345)               |      1.14 |            11.56 |            10.16 |         |
| RabbitMQ.Benchmarks.PrimitivesLonglong.LonglongRead                        |      1.12 |            10.89 |             9.71 |         |
| RabbitMQ.Benchmarks.PrimitivesTimestamp.TimestampRead                      |      1.12 |            10.86 |             9.68 |         |
| RabbitMQ.Benchmarks.PrimitivesBool.BoolWrite5(param1: True, param2: False) |      1.12 |            11.08 |             9.91 |         |
| RabbitMQ.Benchmarks.PrimitivesShort.ShortRead                              |      1.12 |            10.88 |             9.72 |         |
| RabbitMQ.Benchmarks.PrimitivesLong.LongRead                                |      1.11 |            10.83 |             9.72 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayWrite                  |      1.11 |            28.11 |            25.26 |         |
| RabbitMQ.Benchmarks.PrimitivesLonglong.LonglongWrite(value: 12345)         |      1.11 |            11.96 |            10.77 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWriteEmpty             |      1.10 |            21.55 |            19.51 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictWrite                   |      1.10 |            44.15 |            40.05 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableReadPopulated          |      1.10 |           881.56 |           804.29 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrReadPopulated   |      1.09 |           326.49 |           298.99 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicDeliverRead                    |      1.09 |            31.94 |            29.26 |         |
| RabbitMQ.Benchmarks.PrimitivesTimestamp.TimestampWrite                     |      1.09 |            12.25 |            11.26 |         |
| RabbitMQ.Benchmarks.PrimitivesDecimal.DecimalWrite                         |      1.08 |            13.04 |            12.08 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWritePopulated         |      1.08 |           146.53 |           135.84 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishWrite            |      1.07 |           200.85 |           187.07 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableReadEmpty              |      1.06 |            19.71 |            18.51 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntRead                     |      1.06 |            21.08 |            19.82 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishWrite                   |      1.05 |            83.87 |            79.72 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntWrite                    |      1.04 |            22.53 |            21.60 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWritePopulated         |      1.04 |           655.88 |           629.94 |         |
| RabbitMQ.Benchmarks.PrimitivesBool.BoolWrite2(param1: True, param2: False) |      1.04 |             9.97 |             9.62 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.StringWrite                 |      1.03 |            82.96 |            80.52 |         |
| RabbitMQ.Benchmarks.PrimitivesDecimal.DecimalRead                          |      1.03 |            12.39 |            12.05 |         |
| RabbitMQ.Benchmarks.DataTypeShortStringSerialization.ShortstrReadPopulated |      1.03 |           178.36 |           173.73 |         |
| RabbitMQ.Benchmarks.MethodBasicProperties.BasicPropertiesWrite             |      1.03 |            78.22 |            76.26 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueWrite       |      1.02 |            43.31 |            42.62 |         |