@stebet 2021-03-09
### .NET Core 3.1 diff:
| Slower                                                                      | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntGetSize                   |      1.39 |             1.83 |             2.55 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueGetSize      |      1.38 |             9.77 |            13.44 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.StringGetSize                |      1.22 |            12.07 |            14.75 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayReadEmpty               |      1.17 |             8.26 |             9.66 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableGetSizeEmpty            |      1.15 |             8.49 |             9.72 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrGetSizePopulated |      1.11 |           161.62 |           179.19 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWriteEmpty              |      1.08 |            11.40 |            12.27 |         |
| RabbitMQ.Benchmarks.MethodBasicProperties.BasicDeliverSize                  |      1.07 |            16.71 |            17.86 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntRead                      |      1.05 |            11.36 |            11.92 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableReadEmpty               |      1.05 |             9.44 |             9.89 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.NullRead                     |      1.05 |             8.36 |             8.74 |         |
| RabbitMQ.Benchmarks.DataTypeShortStringSerialization.ShortstrWritePopulated |      1.03 |            26.31 |            27.09 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueWrite        |      1.03 |            19.38 |            19.94 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWritePopulated          |      1.03 |           328.78 |           338.25 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.StringWrite                  |      1.02 |            18.46 |            18.90 |         |
| RabbitMQ.Benchmarks.MethodBasicProperties.BasicPropertiesWrite              |      1.02 |            35.93 |            36.60 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishWrite                    |      1.02 |            32.82 |            33.32 |         |

| Faster                                                                    | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayGetSizeEmpty          |      2.57 |             2.15 |             0.83 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrGetSizeEmpty   |      2.11 |             1.86 |             0.88 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrReadEmpty      |      2.06 |             3.22 |             1.56 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictGetSize                |      1.85 |            13.86 |             7.48 |         |
| RabbitMQ.Benchmarks.DataTypeShortStringSerialization.ShortstrGetSizeEmpty |      1.78 |             0.88 |             0.50 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckWrite                          |      1.56 |             3.29 |             2.11 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseWrite                  |      1.42 |             7.09 |             4.98 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckRead                           |      1.39 |             4.60 |             3.31 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueRead       |      1.36 |            23.88 |            17.57 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictRead                   |      1.35 |            20.38 |            15.05 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayWrite                 |      1.26 |            10.29 |             8.16 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishMemoryWrite            |      1.26 |            12.46 |             9.92 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayGetSizePopulated      |      1.26 |            32.29 |            25.73 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicAck.BasicAckWrite                   |      1.25 |            18.39 |            14.70 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseRead                   |      1.24 |             8.26 |             6.66 |         |
| RabbitMQ.Benchmarks.MethodFramingChannelClose.ChannelCloseWrite           |      1.22 |            22.30 |            18.24 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayReadPopulated         |      1.17 |            79.94 |            68.46 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.NullWrite                  |      1.17 |             4.19 |             3.60 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWriteEmpty            |      1.16 |             4.28 |             3.70 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictWrite                  |      1.14 |            20.33 |            17.89 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishMemoryWrite     |      1.14 |            53.24 |            46.90 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableGetSizePopulated      |      1.12 |           222.96 |           199.03 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayGetSize               |      1.11 |             4.73 |             4.26 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicDeliverRead                   |      1.11 |            12.51 |            11.28 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWritePopulated        |      1.10 |            68.35 |            62.06 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableReadPopulated         |      1.07 |           704.85 |           661.16 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.StringRead                 |      1.05 |            22.60 |            21.43 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishWrite           |      1.05 |            96.07 |            91.63 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayRead                  |      1.05 |            11.04 |            10.56 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrReadPopulated  |      1.03 |           219.43 |           212.08 |         |